Rakiba

+++ psh-server/controllers/user.js
@@ -10,6 +10,7 @@ import config from "../config/index.js";
 import catchAsync from "../utils/catchAsync.js";
 import { userServices } from "../services/user.service.js";
 import responseSend from "../utils/responseSend.js";
+import mongoose from "mongoose";
 
 
 export const createUser = async (req, res) => {
@@ -225,6 +226,9 @@ export const updateUser = async (req, res) => {
   try {
     const { id } = req.params;
     // Status Update
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
 
     if (req?.body?.userStatus) {
       await User.updateOne(
@@ -423,6 +427,9 @@ export const verifyOtp = async (req, res) => {
 export const resetPassword = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
     const { newPassword } = req.body;
     const hashedPassword = await bcrypt.hash(
       newPassword,
@@ -443,6 +450,9 @@ export const resetPassword = async (req, res) => {
 
 export const deleteUser = async (req, res, next) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
     await User.findByIdAndDelete(req.params.id);
     res.status(200).json("deleted successfully");
   } catch (err) {
@@ -451,6 +461,10 @@ export const deleteUser = async (req, res, next) => {
 };
 
 export const getUser = catchAsync(async(req, res, next)=>{
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(400).json({ message: "Invalid user ID" });
+  }
+  
   const result = await userServices.getUserById(req.params.id);
   // const encrypted = encrypt(result)
 
@@ -556,6 +570,10 @@ export const getJWT = async (req, res, next) => {
 //update admin user
 export const updateUserAdmin = async (req, res, next) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
+    
     const banner = await User.findByIdAndUpdate(
       req.params.id,
       { $set: req.body },
